refactor custom icons
