+++ frontend/src/layout/navigation/TopNavigation.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import './Navigation.scss'
 import { useActions, useValues } from 'kea'
 import { navigationLogic } from './navigationLogic'
-import { IconBuilding, IconMenu } from './icons'
+import { IconBuilding, IconMenu } from 'lib/components/icons'
 import { userLogic } from 'scenes/userLogic'
 import { Badge } from 'lib/components/Badge'
 import { ChangelogModal } from '~/layout/ChangelogModal'
